moved getHeroDataFromApi on a function file
